move highscore into its own source folder
