+++ src/Etterna/Models/HighScore/HighScore.h
@@ -1,10 +1,10 @@
 #ifndef HIGH_SCORE_H
 #define HIGH_SCORE_H
 
-#include "DateTime.h"
-#include "GameConstantsAndTypes.h"
-#include "Grade.h"
-#include "NoteTypes.h"
+#include "Etterna/Models/Misc/DateTime.h"
+#include "Etterna/Models/Misc/GameConstantsAndTypes.h"
+#include "Etterna/Models/Misc/Grade.h"
+#include "Etterna/Models/Misc/NoteTypes.h"
 #include "RageUtil/Utils/RageUtil_AutoPtr.h"
 
 class XNode;
