+++ src/Etterna/Screen/Others/ScreenSelectMusic.h
@@ -11,7 +11,7 @@
 #include "Etterna/Actor/Base/Sprite.h"
 #include "Etterna/Models/Misc/ThemeMetric.h"
 #include "Etterna/Models/Misc/TimingData.h"
-#include "Etterna/Models/Misc/HighScore.h"
+#include "Etterna/Models/HighScore/HighScore.h"
 #include "Etterna/Models/NoteData/NoteData.h"
 #include "Etterna/Actor/Gameplay/Player.h"
 #include "Etterna/Actor/Gameplay/NoteField.h"
